Adding Current Application Screenshot Walkthrough
Screenshots represent current Volunteer Reporting Form and the screens
with questions being asked of the Volunteer. This is completed using a
downloaded Silverlight application. This is what we are attempting to
simplify with a mobile app, so that it's easier to fill out. Eventually,
this could be replaced by a mobile solution.